backend/pyproject.toml
[project]
name = "fabric-backend"
version = "0.1.0"
description = "Fabric ERP backend — FastAPI"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "asyncpg>=0.29",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "python-dotenv>=1.0",
  "bcrypt>=4.1",
  "pyjwt>=2.8",
  "pyotp>=2.9",
  "redis>=5.0",
  "structlog>=24.1",
  "sentry-sdk[fastapi]>=2.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "S", "SIM", "RUF"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
